+++ app/models.py
@@ -1,15 +1,19 @@
-from sqlalchemy import Column, Integer
+from flask import g, current_app
+import yaml
+
+def init_app(app):
+    if 'models_info' not in g:
+        models_file = current_app.config['MODELS_FILE']
+        with open(models_file) as fp:
+            g.models_info = yaml.safe_load(fp)
+    return g.metadata
 
 
-from app.database import Base
+class Model():
+    def __init__(self, app, name):
+        metadata_fileapp.config
+
+class Dataset(name):
+    def __init__(self):
 
 
-class Segment(Base):
-    __tablename__ = 'segments'
-    length = 3
-    id = Column(Integer, primary_key=True, index=True)
-    track = Column(Integer, index=True)
-    position = Column(Integer)
-
-    def __repr__(self):
-        return f'{self.track}:{self.position * self.length}:{(self.position + 1) * self.length}'