+++ app/processing/__init__.py
@@ -1,0 +1,5 @@
+from .extract import extract
+
+
+def init_app(app):
+    app.cli.add_command(extract)