+++ CHANGELOG.md
@@ -4,12 +4,27 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Latest
+
+### Added
+- Annoy indexes for embeddings storage
+- SQLite database for
+  - Mapping between Annoy ids and `track_id` with timestamps
+  - Track metadata (artist, album, tags)
+
+### Changed
+- Restructured app to have scripts as part of Flask app with `click` instead of `argparse`
+
+### Removed
+- Old processing code using MusiCNN package directly in favor of newer Essentia extractors
+- Support for seaborn plotting
+
 ## [0.2.2] - 2020-07-17
 
-## Added
+### Added
 - Show artist name and track name with link to Jamendo website when playing audio
 
-## Changed
+### Changed
 - Better alerts
 - Merged code for showing segments and trajectories